refactor: types 파일 분리
