+++ src/domain/Restaurant.ts
@@ -1,14 +1,4 @@
-type Category = "한식" | "중식" | "일식" | "아시안" | "양식" | "기타";
-
-type Distance = 5 | 10 | 15 | 20 | 30;
-
-interface RestaurantInfo {
-  name: string;
-  category: Category;
-  distance: Distance;
-  description?: string;
-  url?: string;
-}
+import "../types/restaurant";
 
 export default class Restaurant {
   #info: RestaurantInfo;
