+++ src/domain/RestaurantList.ts
@@ -1,16 +1,5 @@
+import "../types/restaurant";
 import Restaurant from "./Restaurant";
-
-type Category = "한식" | "중식" | "일식" | "아시안" | "양식" | "기타";
-
-type Distance = 5 | 10 | 15 | 20 | 30;
-
-interface RestaurantInfo {
-  name: string;
-  category: Category;
-  distance: Distance;
-  description?: string;
-  url?: string;
-}
 
 export default class RestaurantList {
   #restaurantList: Restaurant[];
